+++ backend/src/common/database/prismaClient.js
@@ -1,0 +1,25 @@
+const { PrismaClient } = require('@prisma/client');
+
+let prismaClient;
+
+function getPrismaClient() {
+  if (!prismaClient) {
+    prismaClient = new PrismaClient();
+  }
+
+  return prismaClient;
+}
+
+async function disconnectDatabase() {
+  if (!prismaClient) {
+    return;
+  }
+
+  await prismaClient.$disconnect();
+  prismaClient = undefined;
+}
+
+module.exports = {
+  getPrismaClient,
+  disconnectDatabase,
+};